Add files via upload
This is our final project for our Data Structures and Algorithms subject. We created a voting system that is inclusive for NU students, the idea of this program is to utilize a data structure that we have learned from the subject. In our case, we utilized ArrayList and Hashmaps for the voting system.